+++ DoorPuzzleSkip/MyPluginInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace DoorPuzzleSkip;
+
+public static class MyPluginInfo
+{
+    public const string PLUGIN_GUID = "DoorPuzzleSkip";
+    public const string PLUGIN_NAME = "Door Puzzle Skip";
+    public const string PLUGIN_VERSION = "1.0.0";
+    public const string PLUGIN_AUTHOR = "Doomnik";
+}